Refactor nested class
