+++ src/main/java/org/mybatis/dynamic/sql/select/UnionQuery.java
@@ -1,0 +1,36 @@
+/*
+ *    Copyright 2016-2023 the original author or authors.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package org.mybatis.dynamic.sql.select;
+
+import java.util.Objects;
+
+public class UnionQuery {
+    private final String connector;
+    private final SelectModel selectModel;
+
+    public UnionQuery(String connector, SelectModel selectModel) {
+        this.connector = Objects.requireNonNull(connector);
+        this.selectModel = Objects.requireNonNull(selectModel);
+    }
+
+    public String connector() {
+        return connector;
+    }
+
+    public SelectModel selectModel() {
+        return selectModel;
+    }
+}
